Update set up instructions, created script for logs, and consolidated scripts
