+++ cron_job.bat
@@ -1,0 +1,2 @@
+python main.py > ./logs/bookings%date:~4,2%%date:~7,2%%date:~10,4%.log
+del *.csv
