+++ logs/log.txt
@@ -1,0 +1,1 @@
+Folder for containing logs!
